completed Feat/fitur 15 annisa

+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/model/DeliveryModel.java
@@ -24,7 +24,7 @@ import java.util.List;
 public class DeliveryModel implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idDelivery;
+    private Long idDelivery;
 
     @NotNull
     @Column(nullable = false)
